Refactor how patches are organized
Patches to the ISO are now batched up and then applied in one pass over
the data. While the performance implications are unclear, it does remove
some duplicated code to locate resources in the PAK files. It also
makes manipulating layer data easer since it moves that logic into the
Patcher object.