Derandomize which intro cinematic is played
Specifically, the FMV that plays between selecting which file to play
and the game actually beginning. Normally, one of three is chose at
random. Since they have different lengths, this can give players a
(small, but still) unfair advantage in a race setting. With this change,
each randomized ISO always uses one of the three. Which one is
determined psuedo-randomly based on the layout.