Fix Gravity Chamber's disappearing stalactites
The stalactites in Gravity Chamber are bugged in the original game such
that they will permanently disappear after they're destroyed the first
time. That would be an issue except that the grapple point doesn't spawn
unless the stalactites are destroyed, thus rendering the item
uncollectable if the stalactites are destroyed once but not immediately
collected.