Fix main menu text rel patch for PAL
This required implementing a mechanism for allowing per-version patches
in the REL. The approach taken here is temporary and should be replaced
by a more permanent linker based approach at a later date. Ideally, the
improved strategy would look more similar to PrimeAPI, where patches are
specified based on the function to intercept rather than the specific
function call instruction to intercept.