Patched Observatory softlock on 1st pass
This occurs when killing pirates before the door is locked at the top of the room when coming from Save Station D or West Tower Entrance
The patch adds a trigger at each of the door at the top.
Known issue with this patch is that the cutscene will play before the 3 pirates before are killed.