Expand c_interface's accepted JSON
Both camcelCase and snake_case names are accepted in most places, but in
general camelCase is what should be used going forward. An expanded
layout format is now supported. This should hopefully be easier to
generate for API consumers and is also future proof. Specifically, we
can add new pickup types (say a freeze trap) or new spawn rooms without
needing to create new variants of the old pseudo-base64 format.